refactor: move SpinalSimVerilatorIoTest
